[FE-163]refactor: 삭제 리팩토링 -삭제 버튼을 누르고 알럿창을 눌러야 삭제가 되야하는데 삭제버튼만 눌러도 삭제되는 버그수정 -알럿 디자인 수정
